Refactoring; split `tests/forms/test_collection` test suite into `tests/forms/test_collection_register` and `tests/forms/test_collection_edit.py`
